Load fontTools installed in Application Support folder
Glyphs allows to install Python modules such as fontTools by clicking
the button in the preferences. They used to be installed in the
following location on the version 2.4.4 or prior:

	/Library/Python/2.7/site-packages

But on Glyphs 2.5.1, they will be install in the following instead:

	~/Library/Application Support/Glyphs/Scripts

When the plugin is loaded from the application, the installed modules
cannot be accessed because the path is not added as a search path. This
patch tries to fix those issues by adding the path just before importing
fontTools.